CAMEL-16928: avoid blocking when resuming from offset repository
Due to the changes in the blocking behavior for Kafka's Consumer
poll(long), this defers the first call of the repository's getState
to when the partition assignment actually happens.
